Prepare for v2 client (#138)
* chore: extract interface

* added new v2 profile

* created dummy client until real dependency is available

* created conditions for the more complex configuration setup

* created v2 client plus its spring boot configuration

* created a yaml for the operate profile

* added new configuration to the autoconfiguration

* adjusted conditional for the operate client v1 configuration

* renamed tests to reflect that they only apply to v1

* v2 operate client is enabled by default

* moved default implementations to interface

* updated readme

* feat: v2 compat

+++ extension/java-client-operate/src/main/java/io/camunda/operate/model/ProcessInstance.java
@@ -2,7 +2,7 @@ package io.camunda.operate.model;
 
 public class ProcessInstance {
   private Long key;
-  private Long processVersion;
+  private Integer processVersion;
   private String bpmnProcessId;
   private Long parentKey;
   private Long parentFlowNodeInstanceKey;
@@ -20,11 +20,11 @@ public class ProcessInstance {
     this.key = key;
   }
 
-  public Long getProcessVersion() {
+  public Integer getProcessVersion() {
     return processVersion;
   }
 
-  public void setProcessVersion(Long processVersion) {
+  public void setProcessVersion(Integer processVersion) {
     this.processVersion = processVersion;
   }
