+++ extension/java-client-operate/src/main/java/io/camunda/operate/search/ProcessInstanceFilterBuilder.java
@@ -16,7 +16,7 @@ public class ProcessInstanceFilterBuilder {
     return this;
   }
 
-  public ProcessInstanceFilterBuilder processVersion(Long processVersion) {
+  public ProcessInstanceFilterBuilder processVersion(Integer processVersion) {
     filter.setProcessVersion(processVersion);
     return this;
   }
